# Removed auto-decoding from the floating point version of std.conv.parse