New version of https://github.com/dlang/phobos/pull/4674

13.5x faster! [Code here](https://gist.github.com/JackStouffer/0cf93d3cc861d699b334e18ea67fb66c#file-test-d).

```
# new
$ ldc2 -O5 -release test.d && ./test
Total time: 387 ms, 506 μs, and 8 hnsecs

# old
$ ldc2 -O5 -release test.d && ./test
Total time: 5 secs, 238 ms, 59 μs, and 4 hnsecs
```