@JackStouffer 2017-01-05
Auto-merge toggled on