@jondegenhardt 2017-01-06
Very nice! This should improve performance of conversions by std.conv.to from string to float or double, right? If that's the case, it might be worth including the std.conv.to benefit in the update notes. Might not be obvious.